// game.server/src/generated/common/enums.ts
export enum GlobalFailCode {
    /**
     * @generated from protobuf enum value: NONE_FAILCODE = 0;
     */
    NONE_FAILCODE = 0,
    /**
     * @generated from protobuf enum value: UNKNOWN_ERROR = 1;
     */
    UNKNOWN_ERROR = 1,
    /**
     * @generated from protobuf enum value: INVALID_REQUEST = 2;
     */
    INVALID_REQUEST = 2,
    /**
     * @generated from protobuf enum value: AUTHENTICATION_FAILED = 3;
     */
    AUTHENTICATION_FAILED = 3,
    /**
     * @generated from protobuf enum value: CREATE_ROOM_FAILED = 4;
     */
    CREATE_ROOM_FAILED = 4,
    /**
     * @generated from protobuf enum value: JOIN_ROOM_FAILED = 5;
     */
    JOIN_ROOM_FAILED = 5,
    /**
     * @generated from protobuf enum value: LEAVE_ROOM_FAILED = 6;
     */
    LEAVE_ROOM_FAILED = 6,
    /**
     * @generated from protobuf enum value: REGISTER_FAILED = 7;
     */
    REGISTER_FAILED = 7,
    /**
     * @generated from protobuf enum value: ROOM_NOT_FOUND = 8;
     */
    ROOM_NOT_FOUND = 8,
    /**
     * @generated from protobuf enum value: CHARACTER_NOT_FOUND = 9;
     */
    CHARACTER_NOT_FOUND = 9,
    /**
     * @generated from protobuf enum value: CHARACTER_STATE_ERROR = 10;
     */
    CHARACTER_STATE_ERROR = 10,
    /**
     * @generated from protobuf enum value: CHARACTER_NO_CARD = 11;
     */
    CHARACTER_NO_CARD = 11,
    /**
     * @generated from protobuf enum value: INVALID_ROOM_STATE = 12;
     */
    INVALID_ROOM_STATE = 12,
    /**
     * @generated from protobuf enum value: NOT_ROOM_OWNER = 13;
     */
    NOT_ROOM_OWNER = 13,
    /**
     * @generated from protobuf enum value: ALREADY_USED_BBANG = 14;
     */
    ALREADY_USED_BBANG = 14,
    /**
     * @generated from protobuf enum value: INVALID_PHASE = 15;
     */
    INVALID_PHASE = 15,
    /**
     * @generated from protobuf enum value: CHARACTER_CONTAINED = 16;
     */
    CHARACTER_CONTAINED = 16
}
// ...
export enum CharacterType {
    /**
     * @generated from protobuf enum value: NONE_CHARACTER = 0;
     */
    NONE_CHARACTER = 0,
    /**
     * 빨강이
     *
     * @generated from protobuf enum value: RED = 1;
     */
    RED = 1,
    /**
     * 상어군
     *
     * @generated from protobuf enum value: SHARK = 3;
     */
    SHARK = 3,
    /**
     * 말랑이
     *
     * @generated from protobuf enum value: MALANG = 5;
     */
    MALANG = 5,
    /**
     * 개굴군
     *
     * @generated from protobuf enum value: FROGGY = 7;
     */
    FROGGY = 7,
    /**
     * 핑크군
     *
     * @generated from protobuf enum value: PINK = 8;
     */
    PINK = 8,
    /**
     * 물안경군
     *
     * @generated from protobuf enum value: SWIM_GLASSES = 9;
     */
    SWIM_GLASSES = 9,
    /**
     * 가면군
     *
     * @generated from protobuf enum value: MASK = 10;
     */
    MASK = 10,
    /**
     * 공룡이
     *
     * @generated from protobuf enum value: DINOSAUR = 12;
     */
    DINOSAUR = 12,
    /**
     * 핑크슬라임
     *
     * @generated from protobuf enum value: PINK_SLIME = 13;
     */
    PINK_SLIME = 13
}
/**
 * @generated from protobuf enum CharacterStateType
 */
export enum CharacterStateType {
    /**
     * @generated from protobuf enum value: NONE_CHARACTER_STATE = 0;
     */
    NONE_CHARACTER_STATE = 0,
    /**
     * 빵야 시전자
     *
     * @generated from protobuf enum value: BBANG_SHOOTER = 1;
     */
    BBANG_SHOOTER = 1,
    /**
     * 빵야 대상 (쉴드 사용가능 상태)
     *
     * @generated from protobuf enum value: BBANG_TARGET = 2;
     */
    BBANG_TARGET = 2,
    /**
     * 현피 중 자신의 턴이 아닐 때
     *
     * @generated from protobuf enum value: DEATH_MATCH_STATE = 3;
     */
    DEATH_MATCH_STATE = 3,
    /**
     * 현피 중 자신의 턴
     *
     * @generated from protobuf enum value: DEATH_MATCH_TURN_STATE = 4;
     */
    DEATH_MATCH_TURN_STATE = 4,
    /**
     * 플리마켓 자신의 턴
     *
     * @generated from protobuf enum value: FLEA_MARKET_TURN = 5;
     */
    FLEA_MARKET_TURN = 5,
    /**
     * 플리마켓 턴 대기 상태
     *
     * @generated from protobuf enum value: FLEA_MARKET_WAIT = 6;
     */
    FLEA_MARKET_WAIT = 6,
    /**
     * 게릴라 시전자
     *
     * @generated from protobuf enum value: GUERRILLA_SHOOTER = 7;
     */
    GUERRILLA_SHOOTER = 7,
    /**
     * 게릴라 대상
     *
     * @generated from protobuf enum value: GUERRILLA_TARGET = 8;
     */
    GUERRILLA_TARGET = 8,
    /**
     * 난사 시전자
     *
     * @generated from protobuf enum value: BIG_BBANG_SHOOTER = 9;
     */
    BIG_BBANG_SHOOTER = 9,
    /**
     * 난사 대상
     *
     * @generated from protobuf enum value: BIG_BBANG_TARGET = 10;
     */
    BIG_BBANG_TARGET = 10,
    /**
     * 흡수 중
     *
     * @generated from protobuf enum value: ABSORBING = 11;
     */
    ABSORBING = 11,
    /**
     * 흡수 대상
     *
     * @generated from protobuf enum value: ABSORB_TARGET = 12;
     */
    ABSORB_TARGET = 12,
    /**
     * 신기루 중
     *
     * @generated from protobuf enum value: HALLUCINATING = 13;
     */
    HALLUCINATING = 13,
    /**
     * 신기루 대상
     *
     * @generated from protobuf enum value: HALLUCINATION_TARGET = 14;
     */
    HALLUCINATION_TARGET = 14,
    /**
     * 감금 중
     *
     * @generated from protobuf enum value: CONTAINED = 15;
     */
    CONTAINED = 15
}
// ...
export enum RoleType {
    /**
     * @generated from protobuf enum value: NONE_ROLE = 0;
     */
    NONE_ROLE = 0,
    /**
     * @generated from protobuf enum value: TARGET = 1;
     */
    TARGET = 1,
    /**
     * @generated from protobuf enum value: BODYGUARD = 2;
     */
    BODYGUARD = 2,
    /**
     * @generated from protobuf enum value: HITMAN = 3;
     */
    HITMAN = 3,
    /**
     * @generated from protobuf enum value: PSYCHOPATH = 4;
     */
    PSYCHOPATH = 4
}
/**
 * @generated from protobuf enum RoomStateType
 */
export enum RoomStateType {
    /**
     * @generated from protobuf enum value: WAIT = 0;
     */
    WAIT = 0,
    /**
     * @generated from protobuf enum value: PREPARE = 1;
     */
    PREPARE = 1,
    /**
     * @generated from protobuf enum value: INGAME = 2;
     */
    INGAME = 2
}